back end
Clone this repo. From within the project folder

Back-end run `npm install`. You 
can then run the server with `node server`. 
The server should run on `localhost` with default port `3000`. If the
server runs on another port for any reason, this is printed to the
console when the server starts, e.g. `Listening on port 3001`.

front-end ,run `npm install`
### Compiles and hot-reloads for development ,npm run serve
Go to `http: // localhost: 8080 /`